requirements.txt
uvicorn==0.32.1
fastapi==0.115.5
google-generativeai==0.8.3
python-dotenv==1.0.1
